+++ src/views/Editor/Thumbnails/index.vue
@@ -414,6 +414,7 @@ const contextmenusThumbnailItem = (): ContextmenuItem[] => {
   position: fixed;
   top: 1vh;
   left: 0;
+  z-index: 1;
 
   &__container {
     padding: 0;
@@ -422,6 +423,10 @@ const contextmenusThumbnailItem = (): ContextmenuItem[] => {
     background-color: white;
     display: flex;
     margin-left:30px;
+    border-radius: 12px;
+  box-shadow: 15px 15px 40px rgba(0, 0, 0, 0.25), 
+              -10px -10px 30px rgba(255, 255, 255, 0.7);
+    
 
     
     .handler-item {