+++ changelog.md
@@ -23,7 +23,7 @@ and this project adheres to([https://semver.org/spec/v2.0.0.html](https://semver
 - Hidden `Aspect Ratio Selector` and `Canvas resolution label` within `right-hand-sidebar`.
 - Changed label within `src/views/Editor/Thumbnails/index.vue` to `Preview (F5)`.
 - Added missing tooltips `src/views/Editor/Thumbnails/index.vue`.
-
+- Added shadow to left-hand-sidebar and increased z-index `src/views/Editor/Thumbnails/index.vue`.
 
 ### Fixed
 -